WIP: factor out Mempool and PollResults
